+++ cloudsmith_cli/cli/commands/resync.py
@@ -7,7 +7,7 @@ from click_spinner import spinner
 from . import main
 from .push import wait_for_package_sync
 from .. import decorators, validators
-from ...core.api.packages import resync_package
+from ...core.api.packages import resync_package as api_resync_package
 from ..exceptions import handle_api_exceptions
 
 
@@ -24,7 +24,7 @@ from ..exceptions import handle_api_exceptions
 @click.pass_context
 def resync(
         ctx, opts, owner_repo_package, skip_errors, wait_interval,
-        no_wait_for_sync):
+        no_wait_for_sync, sync_attempts):
     """
     Resynchronise a package in a repository.
 
@@ -42,6 +42,23 @@ def resync(
     """
     owner, source, slug = owner_repo_package
 
+    resync_package(
+        ctx=ctx, opts=opts, owner=owner, repo=source, slug=slug,
+        skip_errors=skip_errors
+    )
+
+    if no_wait_for_sync:
+        return
+
+    wait_for_package_sync(
+        ctx=ctx, opts=opts, owner=owner, repo=source, slug=slug,
+        wait_interval=wait_interval, skip_errors=skip_errors,
+        attempts=sync_attempts
+    )
+
+
+def resync_package(ctx, opts, owner, repo, slug, skip_errors):
+    """Resynchronise a package."""
     click.echo(
         'Resynchonising the %(slug)s package ... ' % {
             'slug': click.style(slug, bold=True)
@@ -52,18 +69,10 @@ def resync(
     with handle_api_exceptions(ctx, opts=opts, context_msg=context_msg,
                                reraise_on_error=skip_errors):
         with spinner():
-            resync_package(
+            api_resync_package(
                 owner=owner,
-                repo=source,
+                repo=repo,
                 slug=slug
             )
 
     click.secho('OK', fg='green')
-
-    if no_wait_for_sync:
-        return
-
-    wait_for_package_sync(
-        ctx=ctx, opts=opts, owner=owner, repo=source, slug=slug,
-        wait_interval=wait_interval, skip_errors=skip_errors
-    )